+++ lib/deprecations.rb
@@ -1,40 +1,14 @@
-
-DeprecationError = Class.new(ScriptError)
-
 module Deprecations
   autoload(:VERSION, "#{__FILE__[/.*(?=\..+$)/]}/version")
-  require_relative 'deprecations/configuration'
   require_relative 'deprecations/extension'
+  require_relative 'deprecations/behavior'
 
-  class << self
-
-    def call(subject, alternative, outdated)
-      case configuration.behavior
-      when :warn
-        warn(subject, alternative, outdated)
-      when :raise
-        throw!(subject, alternative)
-      end
-      self
-    end
-
-    private
-
-    def throw!(subject, alternative)
-      msg = "`#{subject}` is deprecated"
-      alternative and msg << " - use #{alternative} instead"
-      ex = DeprecationError.new(msg)
-      ex.set_backtrace(caller(3))
-      raise(ex)
-    end
-
-    def warn(subject, alternative, outdated)
-      location = ::Kernel.caller_locations(3,1).last and location = "#{location.path}:#{location.lineno}: "
-      msg = "#{location}[DEPRECATION] `#{subject}` is deprecated"
-      msg << (outdated ? " and will be outdated #{outdated}." : '.')
-      alternative and msg << " Please use `#{alternative}` instead."
-      ::Kernel.warn(msg)
-    end
-
+  def self.call(subject, alternative, outdated)
+    @behavior.call(subject, alternative, outdated)
+    self
   end
+
+  self.behavior = :warn
 end
+
+DeprecationError = Class.new(ScriptError)