+++ lib/deprecations/extension.rb
@@ -14,10 +14,11 @@ module Deprecations
       end
 
       def __method_deprecated!(method, alternative, outdated)
+        defining_context = self
         define_method(method.name) do |*a, &b|
-          decorated = Class === self ? "#{self}." : "#{self.class}#"
+          decorated = Class === self ? "#{self}." : "#{defining_context}#"
           Deprecations.call(
-            "#{decorated}#{__method__}",
+            "#{decorated}#{::Kernel.__method__}",
             UnboundMethod === alternative ? "#{decorated}#{alternative.name}" : alternative,
             outdated
           )
@@ -29,6 +30,9 @@ module Deprecations
         raise(NameError, "undefined method `#{method_name}` for class `#{self}`")
       end
 
+      def __method_alternative(alternative)
+        Symbol === alternative ? (__method(alternative) or __method_not_found!(alternative)) : alternative
+      end
     end
 
     module ClassMethods
@@ -36,9 +40,11 @@ module Deprecations
       include Helper
 
       def deprecated(method_name, alternative = nil, outdated = nil)
-        m = __method(method_name) or __method_not_found!(method_name)
-        a = Symbol === alternative ? (__method(alternative) or __method_not_found!(alternative)) : alternative
-        __method_deprecated!(m, a, outdated)
+        __method_deprecated!(
+          (__method(method_name) or __method_not_found!(method_name)),
+          __method_alternative(alternative),
+          outdated
+        )
       end
     end
 
@@ -48,8 +54,7 @@ module Deprecations
 
       def deprecated(method_name, alternative = nil, outdated = nil)
         m = __method(method_name) or return singleton_class.send(:deprecated, method_name, alternative, outdated)
-        a = Symbol === alternative ? (__method(alternative) or __method_not_found!(alternative)) : alternative
-        __method_deprecated!(m, a, outdated)
+        __method_deprecated!(m, __method_alternative(alternative), outdated)
       end
 
       def deprecated!(alternative = nil, outdated = nil)