+++ spec/spec_helper.rb
@@ -1,4 +1,3 @@
-# configure RSpec
 RSpec.configure do |config|
   config.raise_errors_for_deprecations!
   config.disable_monkey_patching!
@@ -6,5 +5,4 @@ RSpec.configure do |config|
   config.expect_with(:rspec){ |c| c.syntax = :expect }
   config.mock_with(:rspec){ |c| c.syntax = :expect }
 end
-
 require 'deprecations'